finished

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,16 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+
+    const authenticatedUserId = req.session.userId;
+
     try{
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     }catch (error){
         next(error);
@@ -16,8 +22,12 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 export const getNote: RequestHandler = async (req, res, next) => {
 
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try{
+
+        assertIsDefined(authenticatedUserId);
+
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id")
         }
@@ -26,6 +36,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
 
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You cannot access this note")
         }
 
         res.status(200).json(note);
@@ -45,13 +59,18 @@ export const createNotes: RequestHandler<unknown, unknown, CreateNoteBody, unkno
 
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try{
+
+        assertIsDefined(authenticatedUserId);
+
         if(!title){
             throw createHttpError(400, "Note must have a title");
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title, 
             text: text,
         });
@@ -76,8 +95,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try{
+
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id")
@@ -91,6 +113,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if(!note){
             throw createHttpError(404, "Note not found")
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You cannot access this note")
         }
 
         note.title = newTitle;
@@ -107,8 +133,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async(req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try{
+
+        assertIsDefined(authenticatedUserId);
 
         if(!mongoose.isValidObjectId(noteId)){
             throw createHttpError(400, "Invalid note id")
@@ -118,6 +147,10 @@ export const deleteNote: RequestHandler = async(req, res, next) => {
 
         if(!note){
             throw createHttpError(404, "Note not found");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You cannot access this note")
         }
 
         await note.deleteOne();
